requirements.txt
Django==4.2
python-dotenv==1.0.0
psycopg2-binary
python-dotenv
django-import-export
Pillow
requests
bitrix24-rest